back/requirements.txt
bioblend
Flask==1.0.2
flask-cors
python-dotenv
